Feature/simplify child streams (#13)
Reduces configuration needed for adding streams
